Perfumes (#645)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

Adds perfumes using the pollutant system ported from Skyrat (stripped of
its atmos logic).
Perfumes have 10 uses and are not refillable at the moment, but they can
be found in specific map locations or bought from the merchant.

Perfumes can be sprayed on every mob and object, giving it a 10 MINUTES
effect that gives a positive stressor to every character in the vicinity
of the target.

![perfume-export2](https://github.com/user-attachments/assets/76f557ef-c9be-4ea7-b9c9-792097eb7add)
<sub>The color changes based on the fragrance used.</sub>

---------

Co-authored-by: Useroth <37159550+Useroth@users.noreply.github.com>