New wrestling moves! Chokehold & Disarm (#1047)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This pull request adds two new wrestling moves to spice up the combat a
little bit: **Disarms** and **Chokeholds**! Hopefully this'll make
taking non-lethal approaches slightly more doable and diversify
wrestling.

- **Chokehold**
This move is pretty easy to do, just simply get an aggresive grab behind
your opponent's neck and you'll successfully put them on a chokehold!
This'll make it harder for them to break your grab and slightly easier
for you to do follow up moves, like tackling or throwing them, (disarms
too!) It's a pretty good move if you are able to maneouver yourself
behind the enemy or take them by surprise, and can give you an advantage
on the fight if done correctly. Good too since you'll be able to knock
them unconcious faster without having to smash their head in and you
won't get stabbed to death since you're behind them and they can't turn
around.

**Tip!**: If you move whilst having someone on a chokehold, make sure of
having fixed eye on, this way they'll keep facing away from you, so they
won't be able to look at you and get stabbed! Good thing if you're using
someone as a shield or taking hostages.

- **Disarms**
This move is slightly harder to do, grab the HAND (not arm) of the
person holding a weapon, you'll notice the new "take" intent, just click
that and you'll try to do a disarm! If you do it successfully you'll
disarm the person and take their weapon away, or make them drop it to
the ground if you're not that skilled. It's not easy to do with everyone
though, people skilled with their weapon or grabbing them with both
hands are trickier and have less of a chance to disarm. So don't try to
go disarming a knight. But it's a good tool to take weapon's away from
people without having to neccesarily chain them or break their arms, so
this'll probably be mostly used by watchmen and the such.

![Captura de pantalla 2024-09-24
211243](https://github.com/user-attachments/assets/80671b01-0dbe-4131-9bd8-609339fc75e9)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Currently wrestling isn't exactly being used a lot, if you want to deal
with an enemy "non-lethally" the options are usually: Break their arms,
Break their legs, Break their skull, Hit them until unconcious, then
heal any injuries later, or chain (if you're lucky)
This should make it slightly easier to deal with people in a non-lethal
and non-harmful way by taking their weapons away and knocking them
unconcious with oxy damage that heals on its own slightly easier.

Besides, now you'll be able to make the stereotypical stealth takedown
from your favourite game!: Dishonored- I mean The last of Us, or was it
RDR2? Wait no it's Hitman. Wait what am i saying it's from Metal Gear...
or was it from Kingdom come? Idk. It's fun to be able to do it, and now
you get a bonus for getting the drop on your enemy.

PD: I'd appreciate it if you could take a look at the code if you have
experience for anything that i can make better. First time doing
something this big and if i can make the code better i'd love to.
Already tested locally, everything should work.