THE ARCHITECT UPDATE: Craftable Guns by Engineers, Keep Architect, Blunderbusses, and more! (#1654)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

**Adds a Blunderbuss!**

It does less damage than the Arquebus from anywhere but point-blank, and
it has pretty high range, but it does something!

https://github.com/user-attachments/assets/9a3974c3-4662-48b3-8ecf-9ee8d1f9c3a5

**Adds a Clockwork Drill!**

It's a one-of-a-kind item unique to the Architect role that lets you
break down walls much faster! It's one-of-a-kind and spawns with the
architect, who is keep-aligned, so it's used for remodeling without
having to sledgehammer a wall for 20 minutes straight! It's noisy as all
hell so everyone will know the moment you even start wielding it.

https://github.com/user-attachments/assets/acd322c7-dc14-46ca-9773-9fed6aa22b9d

**Replaces the Archivist with the Architect!**

Originally it was going to be its own role, but community members
suggested replacing the Archivist since the Archivist is currently
'boring as hell', and the Architect also fills the role of 'smart guy in
keep that isn't a magician/doctor'. He still has the same duties as the
Archivist when it comes to managing the archives, which let's be honest
no one ever does, but his role no longer sucks so much ass because he
has a secondary job: Being the keep's craftsman, helping with repairs,
creation, and remodeling of the fortress. He spawns with the
aforementioned drill, and the blunderbuss, and gets 2 gun skill.

**Guns, gun parts, and ammo are now craftable!**

Guns can be crafted by those with a 4 in engineering!
Currently, this is only artificers and architects unless you roll
legendary mason. Also, Artificers get 2 gun skill now, since they're
smart guys who aren't going to blow themselves up when trying to use
machinery. They're engineers, for Astrata's sake!

It requires you to make a gun stock, a gun lock, and a gun barrel, and
then combine them to make either an arquebus or blunderbuss. Also,
anyone can saw off an arquebus to make a handgonne, since it's just a
worse short arquebus. You still need to wield it in two hands for some
reason even. Runelock and the arquebus pistol are not craftable as they
are special items.

You can also make a powderflask from the same stuff needed to make a
waterflask plus syrum of ash, replacement ramrods for 1 iron per 3, and
musketballs/ammunition from iron ingots.

You can't make replacement ramrods, powderflasks or musketballs, which
is a huge oversight.

You can now make more guns, but they're pretty rare, only from an
artificer who is willing to put a fair amount of work in!

The Archivist was a boring as fuck role who was only played by people
with a weird 40-year-old-virgin fetish so now it doesn't suck so hard
and is actually fun to play.

You can ACTUALLY remodel the keep now without it taking a whole HOUR to
knock down ONE set of walls.

Blunderbusses are COOL/

---------

Co-authored-by: oilysnake <63020759+oilysnake@users.noreply.github.com>