Shield block change (#1879)
This PR simply removes the need to 'charge up' your block for it to
fully work against projectiles.
So why do this?

Firstly, with the addition of guns that can instantly land a shot on you
with no charge up or travel time, you need faster block mechanics.

Secondly, and the reason I really want this, is because you cannot
roleplay while blocking with a shield. If you are holding down buttons
to block, you cannot type and talk as the same time without opening
yourself up to being shot.

Thirdly, it's confusing as most people don't even realise you need to
charge it up for it to do anything. If I am holding a shield in my
active hand, and using block intent, that should be enough to block.
Anything further is just unintuitive.

Rebalance, more chance to RP, less confusing mechanics.