FIX: Cutlery can't be made. (#1941)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

I goofed up on the code of the last PR by forgetting to include the
blacksmithing skill variable. As a result, Pewter-ware couldn't be made.
This is fixed now and should be working.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Pewter ware can now properly be made by a smith.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->